requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic<2.0.0
python-docx==1.1.0
PyMuPDF==1.23.25
jinja2
python-multipart
